requirements.txt
attrs==22.1.0
black==22.10.0
click==8.1.3
exceptiongroup==1.0.4
iniconfig==1.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.2
pigpio==1.78
pkg_resources==0.0.0
platformdirs==2.5.4
pluggy==1.0.0
pyparsing==3.0.9
pytest==7.2.0
python-dotenv==0.21.0
slack-sdk==3.19.4
tomli==2.0.1
typing_extensions==4.4.0
